requirements.txt
langchain~=0.3.7
langchain-google-vertexai~=2.0.28
pdfplumber~=0.11.7
PyPDF2~=3.0.0
python-docx~=1.2.0
pandas~=2.3.1
openpyxl~=3.1.5
genanki~=0.13.1

h11~=0.16.0
numpy~=1.26.4
pillow~=11.3.0
pip~=25.1
pathlib~=1.0.1
configparser~=7.2.0
attrs~=25.3.0
protobuf~=5.29.5
cryptography~=45.0.5
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
cffi~=1.17.1
typing_extensions~=4.14.1
lxml~=6.0.0
PyMuPDF~=1.26.3
grpcio~=1.74.0
beautifulsoup4~=4.13.4
pytz~=2025.2
requests~=2.32.4
PyYAML~=6.0.2
sniffio~=1.3.1
click~=8.2.1
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
charset-normalizer~=3.4.2
proto-plus~=1.26.1
aiohttp~=3.12.15
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.8.3
urllib3~=2.5.0
packaging~=24.2
google-api-python-client~=2.177.0
google-crc32c~=1.7.1
pyarrow~=19.0.1
shapely~=2.1.1
python-dateutil~=2.9.0.post0
tqdm~=4.67.1
google-cloud-aiplatform~=1.106.0
pydantic~=2.11.7
websockets~=15.0.1
anyio~=4.10.0
tenacity~=8.5.0
multidict~=6.6.3
grpcio-status~=1.62.3
httplib2~=0.22.0
orjson~=3.11.1
SQLAlchemy~=2.0.42
scipy~=1.16.1
numexpr~=2.11.0
cached-property~=2.0.1
chevron~=0.14.0
tzdata~=2025.2
six~=1.17.0
pyparsing~=3.2.3
et_xmlfile~=2.0.0
pydantic_core~=2.33.2
annotated-types~=0.7.0
simplejson~=3.20.1
langchain-core~=0.3.72
langsmith~=0.4.11
langchain-text-splitters~=0.3.8
requests-toolbelt~=1.0.0
pycparser~=2.22
pypdfium2~=4.30.0
Bottleneck~=1.5.0
frozendict~=2.4.6
validators~=0.35.0
uritemplate~=4.2.0
jsonpointer~=3.0.0
jsonpatch~=1.33
google-auth-httplib2~=0.2.0
docstring_parser~=0.17.0
httpx-sse~=0.4.1
pytest~=8.4.1
soupsieve~=2.7
typing_extensions~=4.14.1
rich~=14.1.0
Pygments~=2.19.2
markdown-it-py~=4.0.0
propcache~=0.3.2
mdurl~=0.1.2
platformdirs~=4.3.8
typer~=0.16.1
shellingham~=1.5.4
pyasn1_modules~=0.4.2
yarl~=1.20.1
aiosignal~=1.4.0
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
et_xmlfile~=2.0.0
pydantic_core~=2.33.2
bcrypt~=4.3.0
docstring_parser~=0.17.0
python-pptx~=0.6.23
